Don't log license ID warning with --no-strict-license
Prior to a4bc76e, disabling strict-license would log a warning but not
exit with a failure, but now logging the warning will exit with a
failure status unless fail-on-warnings is also disabled.

This change downgrades license ID failures to warnings all of the time
(since it's a recommendation in the docs anyway) and disables them
entirely when strict-license is disabled, which better reflects what the
flag is useful for - skipping the check.

Now disabling strict-license will not cause a failure exit status and
the default behaviour is still to exit with a failure because of the
warning.

Fixes #69