Make metadata-json-lint dependent on the puppet gem
Using the puppet gem's vendored semantic_puppet should fix the
compatibility and usability issues we're seeing with puppet 5.0.0
with minimal impact to what the user has to do.

Removed the depenency on semantic_puppet everywhere and replaced it
with the puppet gem (which includes the vendored semantic_puppet
libraries).

version_requirement.rb now simply requires 'puppet'