Make SemanticPuppet completely optional and remove dependency on Puppet
This commit changes the two validations that currently use SemanticPuppet
to only use that library if it is already available on the load path. If
not, it falls back to less robust built-in alternatives that should
still be good enough for 99% of use cases. This should eliminate any
depdendency headaches around various versions of Puppet and SemanticPuppet.